pyproject.toml
[tool.poetry]
name = "CMS"
version = "0.1.0"
description = ""
authors = ["wangdong"]

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
djangorestframework = "^3.12.1"
mysqlclient = "^2.0.1"
celery = "^5.0.1"
django-environ = "^0.4.5"
supervisor = "^4.2.1"
uwsgi = "^2.0.19"
httpx = "^0.16.1"
xlwt = "^1.3.0"
openpyxl = "^3.0.5"
reportlab = "^3.5.55"
django-guardian = "^2.3.0"
django-python3-ldap = "^0.11.3"
djangorestframework-simplejwt = "^4.5.0"
django-cid = "^2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
